requirements.txt
algotom==1.7.0
dearpygui==2.1.1
matplotlib==3.10.8
numpy==2.4.0
open3d==0.19.0
opencv_python==4.12.0.88
plotly==6.5.0
pytigre==3.0.0
pyvista==0.46.4
PyYAML==6.0.3
scipy==1.16.3
tqdm==4.67.1
